Disable "AlmostSimpleHint" for L2(31)
Workaround issue #6; the test input there now does not run into an
infinite loop anymore.

However, I am not quite sure if this is the end of the story. Was the
hint simply wrong? Or was it right, but we were wrong to let the hint
goad us into an infinite loop? Or should the system not have generated
that hint in the first place?

And why was that hint added in the first place? It seems to be the only
one for a non-sporadic group. Sadly, it apparently was added before the
original authors used a version control system.