Fix RECOG.IsThisSL2Natural for GL(2,q) with q <= 32
Previously, the code for PSL2 was not applied to the kernel of
the determinant map mod scalars, since there was a extra code path
for q <= 32 which did not take into account that we mod out scalars
only *implicitly*. This is now fixed.

Moreover, I lowered the threshold for the special case from 32 to 11,
as on my machine, the generic code actually performed better. Reducing
it further is not possible without more work (see comments in code).

Fixes #9.