Add checks to diagonal and triangular block code
Without these, we can run into errors, e.g. if we compute random
generators of a kernel, but not enough to generate the full kernel;
in that case, we might think the the kernel is block diagonal,
resp. has block lower triangular shape, when it actually is not.
This could then later on lead to errors.