Remove blackbox code
This code did nothing. It is also clear in which kind of situation
it would ever be called.

In any case, we can easily add it back should we ever come into a
situation where it seems useful, but for now, it just distracts
people who try to learn how recog works.