Disable some secretly broken tests
Move some tests from tst/working to tst/broken; these were secretely
broken; adding a validation check to the TrivialGroup recognition
reveals the brokenness. In one of the tests, the given validation size
clearly was wrong; it is possible that this was/is also the case for the
other two tests.