Improve and doc RandomElm, RandomElmOrd, ...
... GetElmOrd, and SetPseudoRandomStamp, EmptyRecognitionInfoRecord

Adds comments to RandomElm, RandomElmOrd, GetElmOrd, SetPseudoRandomStamp
and EmptyRecognitionInfoRecord documenting the variables involved in
storing and retrieving random elements and their orders.

Restructures `RandomElm`, `RandomElmOrd`, and `GetElmOrd` a little.

Removes the check that `r.nr` must be positive from an if-branch in
`GetElmOrd`.  The input record `r` is always generated by `RandomElm`.
That function always sets `r.nr` to a positive number.