Improve test for FiniteDeletedPermutationModule
Previously, executing this test file would test recognition of finite
deleted permutation modules with a random combination of arguments.
Now, it systematically tests a fixed set of these combinations.

Also now avoid the spurious error that the random generating set we
choose only generated a proper subgroup of the one we wanted to test.