Add mandarins
Old commit msg of a commit which was squashed into this one and which
was nonsense:
RecogniseGeneric: recognise image only once

This is preparation for the implementation of Mandarins. The following
case is going to be handled by the Mandarins. Previously, if in an image
node writing of random elements as SLPs in the image node's nice
generators failed, that image node would be discarded and recomputed.